% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manh.R
\name{qq_manh}
\alias{qq_manh}
\title{Manhatten plots}
\usage{
qq_manh(tab, cols = c(snp = "MarkerName"), p_thr = 5e-08, lab_chr = NULL)
}
\description{
Manhatten plots
}
\examples{
tab <- read_delim("~/Data/SCC/GWAS_processed/gwas/hm3.Consortium_SCC_Meta_SNPonly_impu_filter_anno.txt", delim = " ")
tab22 <- filter(tab, Chr == 22)

}
